feat: 가이드 단계 추가
